requirements.txt
pytest==7.1.3
pytest-html==3.1.1
pytest-metadata==2.0.2
attrs==22.1.0
more-itertools==9.0.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
requests==2.28.1
jsonpath==0.82
pytest-html-reporter~=0.2.3
pytest-json==0.4.0
invoke==1.7.3
allure-pytest==2.11.1
pytest-sugar==0.9.5
pytest-rerunfailures==10.2
pytest-parallel==0.1.1
pytest-xdist==2.5.0
six==1.16.0
PyYAML==6.0